backend/requirements.txt
fastapi
uvicorn[standard]
httpx
langchain-google-genai
google-genai
anyio
langchain